Use global interval for asp.net duration trigger (#967)
* Use global interval for asp.net duration trigger

* Schema update

* PR feedback

* Update dependencies from https://github.com/dotnet/diagnostics build 20211007.1

Microsoft.Diagnostics.Monitoring , Microsoft.Diagnostics.Monitoring.EventPipe
 From Version 5.0.0-preview.21506.1 -> To Version 5.0.0-preview.21507.1

+++ src/Microsoft.Diagnostics.Monitoring.WebApi/Utilities.cs
@@ -62,7 +62,7 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi
             return scope;
         }
 
-        public static MonitoringSourceConfiguration GetTraceConfiguration(Models.TraceProfile profile, int metricsIntervalSeconds)
+        public static MonitoringSourceConfiguration GetTraceConfiguration(Models.TraceProfile profile, float metricsIntervalSeconds)
         {
             var configurations = new List<MonitoringSourceConfiguration>();
             if (profile.HasFlag(Models.TraceProfile.Cpu))
